+++ packages/cds-migrator/src/migrations/q42023/index.ts
@@ -5,6 +5,7 @@ import catchAllPropMigrations from './catchAllPropMigrations';
 import componentMigrations from './componentMigrations';
 import manualPropMigrations from './manualPropMigrations';
 import propMigrations from './propMigrations';
+import propToAttributeValueMigrations from './propToAttributeValueMigrations';
 import removedComponents from './removedComponents';
 import removedPropMigrations from './removedPropMigrations';
 
@@ -16,4 +17,5 @@ export default async function main(tree: Tree) {
   await removedPropMigrations(tree);
   await manualPropMigrations(tree);
   await attributeValueToBooleanMigrations(tree);
+  await propToAttributeValueMigrations(tree);
 }