+++ packages/mobile/CHANGELOG.md
@@ -8,6 +8,12 @@ All notable changes to this project will be documented in this file!
 
 <!-- template-start -->
 
+## 5.13.0 (9/27/2023 PST)
+
+#### 🚀 Updates
+
+- [DX-2337] chore: deprecated vertical prop. [[#2284](https://github.cbhq.net/frontend/cds/pull/2284)] [[DX-2337](https://jira.coinbase-corp.com/browse/DX-2337)]
+
 ## 5.12.2 ((9/27/2023, 10:09 AM PST))
 
 This is an artificial version bump with no new change.